src/commons/requirements.txt
aws-cdk-lib==2.154.1
constructs>=10.0.0,<11.0.0
toml==0.10.2
pydantic==2.8.2
